requirements.txt
strands-agents
strands-agents-tools
boto3>=1.42.3
botocore>=1.42.3
bedrock-agentcore==1.1.1
bedrock-agentcore-starter-toolkit==0.2.3
aws-opentelemetry-distro==0.14.0
pyyaml
moto[all]>=5.0.0
pytest
pytest-asyncio
